refactor(comp:switch): style and demo
