+++ packages/components/checkbox/style/index.less
@@ -179,9 +179,6 @@
   &.@{checkbox-prefix}-focused:not(.@{checkbox-prefix}-disabled) {
     color: @checkbox-focus-color;
     border-color: @checkbox-focus-color;
-    & + .@{checkbox-prefix} {
-      border-left-color: @checkbox-focus-color;
-    }
   }
 
   .@{checkbox-prefix} {
