+++ packages/components/input/style/mixin.less
@@ -158,7 +158,8 @@
     background-color: @input-addon-bg-color;
     border: @input-border-width @input-border-style @input-border-color;
     border-radius: @input-border-radius;
-
+    white-space: nowrap;
+    
     &:first-child {
       border-right: 0;
       border-top-right-radius: 0;
