+++ packages/components/switch/demo/Basic.md
@@ -7,12 +7,8 @@ order: 0
 
 ## zh
 
-* 通过v-model:checked进行双向数据绑定
-* 父组件改变checked的状态，子组件监听到checked的变化，作出相应改变
-* 子组件通过click事件改变checked的状态，向外emit触发change事件
+通过v-model:checked进行双向数据绑定
 
 ## en
 
-* Two-way data binding through v-model:checked.
-* The parent component changes the checked state, and the child component listens to the checked changes and makes corresponding changes.
-* The child component changes the checked state through click, and triggers change and click events outward.
+Two-way data binding through v-model:checked.
