+++ packages/components/switch/demo/Basic.vue
@@ -1,7 +1,8 @@
 <template>
-  <IxSwitch v-model:checked="checked" @change="updateChecked"></IxSwitch>
-  <div class="mt-5">checked: {{ checked }}</div>
-  <IxButton class="mt-5" @click="changeChecked">父组件触发改变</IxButton>
+  <IxSpace direction="vertical" :size="20">
+    <IxSwitch v-model:checked="checked" @change="updateChecked"></IxSwitch>
+    <IxButton @click="changeChecked">Change</IxButton>
+  </IxSpace>
 </template>
 <script lang="ts">
 import { defineComponent, ref } from 'vue'
@@ -25,8 +26,3 @@ export default defineComponent({
   },
 })
 </script>
-<style lang="less" scoped>
-.mt-5 {
-  margin-top: 5px;
-}
-</style>
