+++ packages/components/switch/demo/Content.md
@@ -1,0 +1,14 @@
+---
+title:
+  zh: 内容
+  en: Content
+order: 4
+---
+
+## zh
+
+向开关中加入内容。
+
+## en
+
+Add content to the `Switch`.
