+++ packages/components/switch/demo/Disabled.md
@@ -2,17 +2,13 @@
 title:
   zh: disabled的使用
   en: disabled usage
-order: 0
+order: 1
 ---
 
 ## zh
 
-* 默认disabled为false
-* disabled为true，则switch不可点，不会向外界触发事件
-* disabled为false，则switch正常可操作
+通过设置 `disabled` 将`Switch`设为禁用状态。
 
 ## en
 
-* The default disabled is false.
-* If disabled is true, the switch cannot be clicked, the state will not change, and events will not be triggered to the outside world.
-* If disabled is false, the switch can be operated normally.
+Set `Switch` to disabled status by setting `disabled`.
