--- packages/components/switch/demo/Focus.md
@@ -1,17 +1,0 @@
----
-title:
-  zh: focus、blur方法的使用
-  en: focus and blur methods usage
-order: 0
----
-
-## zh
-
-* 外部调用focus()，switch会获得焦点
-* 外部调用blur()，switch会失去焦点
-* autofocus，一打开页面即自动获得焦点
-
-## en
-
-* Call focus() externally, switch will get the focus.
-* When blur() is called externally, the switch will lose focus.
