--- packages/components/switch/demo/Inner.md
@@ -1,18 +1,0 @@
----
-title:
-  zh: 内部props、slots的使用
-  en: Inner props and slots usage
-order: 0
----
-
-## zh
-
-* 只配置props，props生效
-* 同时配置props、slots，slots生效
-* 可以配置disabled
-
-## en
-
-* Only configure props, props take effect.
-* Also configure Props Slots, Slots works.
-* Can be configured disabled.
