+++ packages/components/switch/demo/Loading.vue
@@ -1,10 +1,8 @@
 <template>
-  <div>
+  <IxSpace direction="vertical" :size="20">
     <IxSwitch v-model:checked="checked1" :loading="loading"></IxSwitch>
-  </div>
-  <div class="mt-5">
     <IxSwitch v-model:checked="checked2" :loading="loading"></IxSwitch>
-  </div>
+  </IxSpace>
 </template>
 <script lang="ts">
 import { defineComponent, ref } from 'vue'
@@ -15,21 +13,11 @@ export default defineComponent({
     const checked2 = ref(false)
     const loading = ref(true)
 
-    const updateChecked = (updateChecked: boolean) => {
-      console.log(updateChecked)
-    }
-
     return {
       checked1,
       checked2,
-      updateChecked,
       loading,
     }
   },
 })
 </script>
-<style lang="less" scoped>
-.mt-5 {
-  margin-top: 5px;
-}
-</style>
